requirements.txt
dotenv>=0.9.9
Flask>=3.1.1
psycopg2-binary>=2.9.10
python-dateutil>=2.9.0.post0
requests>=2.32.4
SQLAlchemy>=2.0.43
gunicorn>=21.2.0